+++ includes/admin/class-wp-statistics-admin-assets.php
@@ -431,8 +431,8 @@ class Admin_Assets
             $screen = get_current_screen();
 
             if (stristr($screen->id, 'wps_')) {
-                wp_enqueue_script('feedbackbird-app-script', 'https://cdn.jsdelivr.net/gh/feedbackbird/assets@master/wp/app.js?uid=01H34YMWXSA9XPS61M4S11RV6Z');
-                wp_add_inline_script('feedbackbird-app-script', sprintf('var feedbackBirdObject = %s;', json_encode([
+                wp_enqueue_script('feedbackbird-widget', 'https://cdn.jsdelivr.net/gh/feedbackbird/assets@master/wp/app.js?uid=01H34YMWXSA9XPS61M4S11RV6Z');
+                wp_add_inline_script('feedbackbird-widget', sprintf('var feedbackBirdObject = %s;', json_encode([
                     'user_email' => function_exists('wp_get_current_user') ? wp_get_current_user()->user_email : '',
                     'platform'   => 'wordpress-admin',
                     'config'     => [
@@ -453,7 +453,7 @@ class Admin_Assets
                 ])));
 
                 add_filter('script_loader_tag', function ($tag, $handle, $src) {
-                    if ('feedbackbird-app-script' === $handle) {
+                    if ('feedbackbird-widget' === $handle) {
                         return preg_replace('/^<script /i', '<script type="module" crossorigin="crossorigin" ', $tag);
                     }
                     return $tag;